getAllContact and getById
