+++ controllers/contacts/addContact.js
@@ -1,12 +1,11 @@
 const { requestError } = require('../../helpers');
-const Book = require('../../models/contact')
+const {Contact} = require('../../models/contact')
 
 
 const add = async (req, res) => {
-  const data = await Book.create(req.body)
+  const data = await Contact.create(req.body)
   if(!data)throw requestError(500,"Server error")
   res.status(201).json(data)
 }
-
 
 module.exports = add;
