+++ controllers/contacts/getById.js
@@ -1,10 +1,12 @@
-const {getContactById} = require('../../models/contacts');
+const {Contact} = require('../../models/contact');
 const {requestError} = require('../../helpers');
 
 const getById = async (req, res) => {
-  const data = await getContactById(req.params.Id)
+  const { Id } = req.params;
+  const data = await Contact.findOne({_id: Id})
   if(!data) throw requestError(404, "Not found");
   res.json(data);
 }
+
 
 module.exports = getById;
