+++ controllers/contacts/index.js
@@ -1,12 +1,12 @@
-// const getAll = require('./getAllContats');
-// const getById = require('./getById');
+const getAll = require('./getAllContats');
+const getById = require('./getById');
 const add = require('./addContact');
 // const updateById = require('./updateById');
 // const removeById = require('./removeById');
 
 module.exports = {
-    // getAll,
-    // getById,
+    getAll,
+    getById,
     add,
     // updateById,
     // removeById
