+++ helpers/index.js
@@ -1,11 +1,9 @@
 const requestError = require('./requestError');
-const generatorId = require('./generatorId');
 const ctrlWrapper = require('./ctrlWrapper')
 const handleSaveErrorrs = require('./handleSaveErrors')
 
 module.exports = {
     requestError,
-    generatorId,
     ctrlWrapper,
     handleSaveErrorrs
 }
