+++ routes/api/contacts.js
@@ -3,19 +3,17 @@ const router = express.Router()
 const ctrl = require('../../controllers/contacts')
 const {ctrlWrapper} = require('../../helpers')
 const { validateBody } = require('../../middlewares')
-const schemas = require('../../schemas/contacts')
+const {schemas} = require('../../models/contact')
 
 
-// router.get('/', ctrlWrapper(ctrl.getAll))
+router.get('/', ctrlWrapper(ctrl.getAll))
 
-// router.get('/:Id', ctrlWrapper(ctrl.getById))
+router.get('/:Id', ctrlWrapper(ctrl.getById))
 
-// router.post('/', validateBody(schemas.schemaAdd), ctrlWrapper(ctrl.add))
-router.post('/', ctrlWrapper(ctrl.add))
+router.post('/', validateBody(schemas.schemaAdd), ctrlWrapper(ctrl.add))
 
 // router.delete('/:Id', ctrlWrapper(ctrl.removeById))
 
 // router.put('/:Id', validateBody(schemas.schemaUpdate), ctrlWrapper(ctrl.updateById))
-
 
 module.exports = router;
